refactor(l5): move source-refs and candidate-dedupe from src/ into server/domain (issue 864)
Two domain leaves move out of src/ into server/domain/, meeting the server/
standard whole. The old src/ paths become L5 shims so every legacy importer
keeps its import line untouched and stays out of the pre-commit gate.

Moves (old -> new):
- src/source-refs.ts        -> server/domain/source-refs.ts
- src/source-refs.test.ts   -> server/domain/source-refs.test.ts
- src/candidate-dedupe.ts   -> server/domain/candidate-dedupe.ts
- src/candidate-dedupe.test.ts -> server/domain/candidate-dedupe.test.ts

Transformations forced by the server/ overrides, no behavior change:
- candidate-dedupe: runCandidateDedupe (152 lines, complexity 15) split by
  responsibility into selectMergePairs, lockPairRows, and applyMerge. The SQL,
  the transaction shape, and the summary accounting are byte-identical.
- source-refs.test.ts: each it body hoisted to a module-scope function called
  by the same test name; three non-null assertions become expectDefined.
- candidate-dedupe.test.ts: fifteen non-null assertions become expectDefined
  from scripts/test-support/expect-defined.ts. Test names, order, and
  assertion meaning unchanged.

Runtime closure of src/ under server/main.ts: 56 -> 54.

distill-window.ts and distiller.ts were in the same lane brief and are NOT in
this commit: distill-window needs a claimDistillBatch split plus roughly
thirty non-null assertions across source and test, which did not fit the lane
timebox with the tests verified. Both remain in src/ untouched.
